app/requirements.txt
Flask==3.0.0
redis==5.0.1
google-cloud-storage==2.14.0
